+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Fetcher Changelog
 
+## [1.4.2](https://github.com/LerianStudio/fetcher/releases/tag/v1.4.2)
+
+- Fixes:
+  - Bumped lib-commons to v5.1.3 to address issues with HTTP/1.1 tmclient transport.
+
+Contributors: @jeffersonrodrigues92, @lerian-studio
+
+[Compare changes](https://github.com/LerianStudio/fetcher/compare/v1.4.1...v1.4.2)
+
+---
+
+## [1.4.1](https://github.com/LerianStudio/fetcher/releases/tag/v1.4.1)
+
+- Fixes:
+  - Fixed issue with subscribing to environment-scoped tenant-events channel.
+
+Contributors: @jeffersonrodrigues92
+
+[Compare changes](https://github.com/LerianStudio/fetcher/compare/v1.4.0...v1.4.1)
+
+---
+
 ## [1.3.0](https://github.com/LerianStudio/fetcher/releases/tag/v1.3.0)
 
 Features: